First cut at machines page
For now just introduce the machinery: we retrieve and convey machines info
when querying for first test page. Two nested templates are built and
sent to browser that splits again the content in 2 parts (machines and tests).

Tests are shown immediately, machines only on clicking the number of workers
in active tests. Machines are shown with a modal window.

All the UI is still to workout.