Replace ellipsis with sha first digits
Following a request on github from xoto:

I have one question, is it possible to include part of the commit ref in
the test name? If people (I am guilty of this!) submit multiple tests with
altered code on the same branch, we need to click on the test to remember
which code version it is. A couple of letters from the commit ref giving
test_name_xy instead of test_name might be a bit easier to follow.

See https://github.com/official-stockfish/Stockfish/issues/1267