Optimize DB query
Retrieve only neded fields, or at least not all fields.

The most expansive are SPSA related and luckily we don't use them. We also
don't load tasks for finished runs, for active runs we need them to
compute machine page, so split the query according if run is finished.

With this patch we pass from loading 10MB(!) for first page request to
500K, following pages now require less than 50K.

This patch has a noticeable impact on response time, because  MongoDB
runs on a remote server and data has to travel the net!

Now that loading pages is much faster, I have increased scrollThreshold
for infinite scrolling, as a result spinner icon does not appear
almost anymore while scrolling.