Further improve DB query performance
Use a pipe to fetch only active tasks for active runs, instead of all the
tasks of active runs (included the ones already finished).

This further reduce the size of the fetched results. As an example,
with 7 active runs, we went from 222K to 95K, cutting in half the data.

Also count active workers at server side, simplifying template processing.