Add filtering for username
Clicking on test author toggles the filtering of tests by selected name. It is
a toggle, so it can be reverted re-clicking or clicking on the filter icon in
the top bar.

Reverting form filtering is done fully on browser side that backups current state
before filtering and reverts it back without requery.