Don't overwrite #machines when filering
In setupGauges we bound resetGauges to #machines, but every time a machine
table is reload, like when filtering, the event binding is dropped and
resetGauges is not called anymore, leaving machines tab open.

The fix is to move #machines in the base template so that it is never
overwritten dynamically.

This fixes issue #3