Compute and show games per minute
Adding games per minute to machine view has been more difficult than expected!

Mainly we proceed as follows:

- Compute finished games for each active task (DB)
- Send this value when different form previous (websocket)
- Collect a long enough series of games counters values (browser)
- Compute GPM as difference between the latest and oldest (browser)

And finally show in machine page.

The approach is not perfect because it assumes that at every update
there is a new game counter for each active task, but this is not
always true, so that we end up with a series of 12 samples and
because websocket tick is of 5 seconds we assume the oldest and newest
are 1 minute apart, but it may take much more to accumulate 12 samples.

On the other hand this is a small error because it affects only very
slow tasks that do not add much in total GPM.